use 64bit numbers, add config to let Lexer use origin string in lua code steam

+++ src/types.rs
@@ -1,6 +1,6 @@
 
-pub type IntType = i32;
-pub type FloatType = f32;
+pub type IntType = i64;
+pub type FloatType = f64;
 pub enum Number {
     Int(IntType),
     Float(FloatType),
